+++ src/modules/client/inspiration/architect-profile/hooks/use-project-item.tsx
@@ -1,17 +1,16 @@
 import { useRouter } from 'expo-router';
-import React, { useCallback, useState } from 'react';
-import type {
-  ListRenderItemInfo,
-  NativeScrollEvent,
-  NativeSyntheticEvent,
-} from 'react-native';
+import React, { useCallback } from 'react';
+import type { ListRenderItemInfo } from 'react-native';
 
+import { useImageSlider } from '@/core';
 import { Image, WIDTH } from '@/shared/components';
 
 import type { ProjectItemProps } from '../../types';
 
 export const useProjectItem = ({ item }: ProjectItemProps) => {
-  const snapToOffsets = item?.projectImages.map((_, index) => index * WIDTH);
+  const { handleScroll, currentIndex, snapToOffsets, totalImages } =
+    useImageSlider(item?.projectImages);
+
   const renderItem = useCallback(
     ({ item }: ListRenderItemInfo<string>) => (
       <Image
@@ -22,13 +21,6 @@ export const useProjectItem = ({ item }: ProjectItemProps) => {
     ),
     []
   );
-  const [currentIndex, setCurrentIndex] = useState(0);
-  const totalImages = item?.projectImages?.length ?? 0;
-  const handleScroll = (event: NativeSyntheticEvent<NativeScrollEvent>) => {
-    const offsetX = event.nativeEvent.contentOffset.x;
-    const index = Math.round(offsetX / WIDTH);
-    setCurrentIndex(index);
-  };
 
   const router = useRouter();
   const navigateToProjectDetails = () => {
