+++ src/modules/client/inspiration/hooks/use-project-item.tsx
@@ -1,31 +1,22 @@
 import { useRouter } from 'expo-router';
-import React, { useCallback, useState } from 'react';
-import type {
-  ListRenderItemInfo,
-  NativeScrollEvent,
-  NativeSyntheticEvent,
-} from 'react-native';
+import React, { useCallback } from 'react';
+import type { ListRenderItemInfo } from 'react-native';
 
-import { Image, WIDTH } from '@/shared/components';
+import { useImageSlider } from '@/core';
+import { Image } from '@/shared/components';
 
-import { projectList } from '../dump-data';
 import type { ProjectItemProps } from '../types';
 
 export const useProjectItem = ({ item }: ProjectItemProps) => {
-  const snapToOffsets = projectList.map((_, index) => index * WIDTH);
+  const { handleScroll, currentIndex, snapToOffsets, totalImages } =
+    useImageSlider(item?.projectImages);
+
   const renderItem = useCallback(
     ({ item }: ListRenderItemInfo<string>) => (
       <Image source={{ uri: item }} className="h-screen w-screen" />
     ),
     []
   );
-  const [currentIndex, setCurrentIndex] = useState(0);
-  const totalImages = item?.projectImages?.length ?? 0;
-  const handleScroll = (event: NativeSyntheticEvent<NativeScrollEvent>) => {
-    const offsetX = event.nativeEvent.contentOffset.x;
-    const index = Math.round(offsetX / WIDTH);
-    setCurrentIndex(index);
-  };
 
   const router = useRouter();
   const navigateToProfile = () => {
